requirements.txt
aiofauna==0.1.73
aiohttp==3.8.5
aiohttp-cors==0.7.0
aiohttp-sse==2.1.0
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
beautifulsoup4==4.12.2
boto3==1.28.21
botocore==1.31.21
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cheapcone==0.1.9
click==8.1.6
decorator==5.1.1
docker==6.1.3
exceptiongroup==1.1.2
fastapi==0.101.0
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
lazy_loader==0.3
librosa==0.10.0.post2
llvmlite==0.40.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msgpack==1.0.5
multidict==6.0.4
numba==0.57.1
numpy==1.24.4
openai==0.27.8
packaging==23.1
pooch==1.6.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.4.0
pydub==0.25.1
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==15.0.0
regex==2023.6.3
requests==2.31.0
rich==13.5.2
s3transfer==0.6.1
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
starlette==0.27.0
threadpoolctl==3.2.0
tiktoken==0.4.0
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
websocket-client==1.6.1
yarl==1.9.2
